Handle inheritance correctly
If fields or methods are not found in the class, pitifulvm will try
to find them from the parent class recursively.

Note that ptitfulvm have not supported objects' fields in the parent
class yet, so `getfield` and `putfield` still don't find fields from
the parent class.

Add a new test script: "Inherit.java"